+++ plugins/woocommerce-admin/packages/components/src/table/index.js
@@ -54,7 +54,7 @@ class TableCard extends Component {
 		this.onClickDownload = this.onClickDownload.bind( this );
 		this.onCompare = this.onCompare.bind( this );
 		this.onPageChange = this.onPageChange.bind( this );
-		this.onSearch = this.onSearch.bind( this );
+		this.onSearchChange = this.onSearchChange.bind( this );
 		this.selectRow = this.selectRow.bind( this );
 		this.selectAllRows = this.selectAllRows.bind( this );
 	}
@@ -128,12 +128,12 @@ class TableCard extends Component {
 					}
 
 					const showCols = without( prevState.showCols, key );
-					onColumnsChange( showCols );
+					onColumnsChange( showCols, key );
 					return { showCols };
 				}
 
 				const showCols = [ ...prevState.showCols, key ];
-				onColumnsChange( showCols );
+				onColumnsChange( showCols, key );
 				return { showCols };
 			} );
 		};
@@ -179,8 +179,13 @@ class TableCard extends Component {
 		}
 	}
 
-	onSearch( values ) {
-		const { compareParam, searchBy, baseSearchQuery } = this.props;
+	onSearchChange( values ) {
+		const {
+			baseSearchQuery,
+			compareParam,
+			onSearch,
+			searchBy,
+		} = this.props;
 		// A comma is used as a separator between search terms, so we want to escape
 		// any comma they contain.
 		const labels = values.map( v => v.label.replace( ',', '%2C' ) );
@@ -197,6 +202,8 @@ class TableCard extends Component {
 				search: undefined,
 			} );
 		}
+
+		onSearch( values );
 	}
 
 	selectAllRows( event ) {
@@ -266,6 +273,7 @@ class TableCard extends Component {
 			isLoading,
 			onClickDownload,
 			onQueryChange,
+			onSort,
 			query,
 			rowHeader,
 			rowsPerPage,
@@ -317,7 +325,7 @@ class TableCard extends Component {
 							allowFreeTextSearch={ true }
 							inlineTags
 							key="search"
-							onChange={ this.onSearch }
+							onChange={ this.onSearchChange }
 							placeholder={ labels.placeholder || __( 'Search by item name', 'woocommerce-admin' ) }
 							selected={ searchedLabels }
 							showClearButton={ true }
@@ -376,7 +384,6 @@ class TableCard extends Component {
 							rowHeader={ rowHeader }
 							caption={ title }
 							query={ query }
-							onSort={ onQueryChange( 'sort' ) }
 						/>
 					</Fragment>
 				) : (
@@ -386,7 +393,7 @@ class TableCard extends Component {
 						rowHeader={ rowHeader }
 						caption={ title }
 						query={ query }
-						onSort={ onQueryChange( 'sort' ) }
+						onSort={ onSort || onQueryChange( 'sort' ) }
 					/>
 				) }
 
@@ -453,6 +460,14 @@ TableCard.propTypes = {
 	 * A function which returns a callback function which is called upon the user changing the visiblity of columns.
 	 */
 	onColumnsChange: PropTypes.func,
+	/**
+	 * A function which is called upon the user searching in the table header.
+	 */
+	onSearch: PropTypes.func,
+	/**
+	 * A function which is called upon the user changing the sorting of the table.
+	 */
+	onSort: PropTypes.func,
 	/**
 	 * Whether the table must be downloadable. If true, the download button will appear.
 	 */
@@ -522,6 +537,8 @@ TableCard.defaultProps = {
 	isLoading: false,
 	onQueryChange: noop,
 	onColumnsChange: noop,
+	onSearch: noop,
+	onSort: undefined,
 	query: {},
 	rowHeader: 0,
 	rows: [],