+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,5 +1,7 @@
 # unreleased
 - SearchListItem component: new `countLabel` prop that will overwrite the `item.count` value.
+- AdvancedFilters component: fire `onAdvancedFilterAction` for match changes.
+- TableCard component: add `onSearch` an `onSort` function props.
 - Add new component `<List />` for displaying interactive list items.
 
 # 3.1.0